+++ src/app/components/calendar/calendar.component.ts
@@ -1,84 +1,33 @@
 import { Component, OnDestroy, OnInit } from '@angular/core';
-import {  Subject } from 'rxjs';
-import { takeUntil } from 'rxjs/operators';
+import { Subject } from 'rxjs';
 import { Reminder } from 'src/app/interfaces/reminder';
 import { CalendarService } from 'src/app/services/calendar.service';
 import { WeatherService } from 'src/app/services/weather.service';
 import { MatDialog } from '@angular/material/dialog';
 import { ReminderFormComponent } from '../reminder-form/reminder-form.component';
 
-//mine
 import * as moment from 'moment';
 
-import { CalendarEvent, CalendarLayoutData, CalendarViewDateType, FormModeData, PersistenceData } from 'src/app/models/calender-models';
+import {
+  CalendarEvent,
+  CalendarLayoutData,
+  FormModeData,
+  PersistenceData,
+} from 'src/app/interfaces/calender-models';
 import { kToF, rangeFor } from 'src/app/helpers/helper-fns';
 import { ConstantService } from 'src/app/services/calender-constants';
 import { PersistenceServiceService } from 'src/app/services/persistence-service.service';
 import * as icons from '@fortawesome/free-solid-svg-icons';
 
-
 @Component({
   selector: 'app-calendar',
   templateUrl: './calendar.component.html',
-  styleUrls: ['./calendar.component.scss']
+  styleUrls: ['./calendar.component.scss'],
 })
 export class CalendarComponent implements OnInit, OnDestroy {
-
   onDestroy$ = new Subject<boolean>();
 
-  constructor(
-    private calendarService: CalendarService,
-
-    //todo marry
-    private weatherService: WeatherService,
-
-    private matDialog: MatDialog,
-    private persistenceService: PersistenceServiceService,
-    private constantService: ConstantService
-  )  {
-    this.now = new Date();
-    this.availableTimesList = this.constantService.AVAILABLE_TIMES_LIST;
-  }
-
-  ngOnInit(): void {
-
-    this.initCal(this.now);
-    this.setLocation();
-
-    //theirs
-    // this.calendarService.list(new Date())
-    //   .pipe(takeUntil(this.onDestroy$))
-    //   .subscribe((reminders: Reminder[]) => {
-    //     reminders.map((reminder: Reminder) => {
-    //       return {
-    //         ...reminder,
-    //         weather: this.getWeather(reminder.city),
-    //       };
-    //     });
-    //     console.log(reminders);
-    //   });
-  }
-
-  getWeather(city: string) {
-    const x = this.weatherService.getWeatherInformation(city);
-    console.log(x);
-    return x;
-  }
-
-  ngOnDestroy() {
-    this.onDestroy$.next(true);
-    this.onDestroy$.complete();
-  }
-
-  openReminderForm(reminder?: Reminder) {
-    this.matDialog.open(ReminderFormComponent, {
-      data: {
-        reminder,
-      },
-    });
-  }
-  //end theres.
-
+  
   range = rangeFor;
   now: Date;
   icons = icons;
@@ -120,10 +69,31 @@ export class CalendarComponent implements OnInit, OnDestroy {
       }
     | undefined;
 
-  dialogResult = '';
-  dialogContainerRef: any;
   hasSelection = false;
 
+  constructor(
+    private calendarService: CalendarService,
+
+    private weatherService: WeatherService,
+
+    private matDialog: MatDialog,
+    private persistenceService: PersistenceServiceService,
+    private constantService: ConstantService
+  ) {
+    this.now = new Date();
+    this.availableTimesList = this.constantService.AVAILABLE_TIMES_LIST;
+  }
+
+  ngOnInit(): void {
+    this.initCal(this.now);
+    this.setLocation();
+
+  }
+
+  ngOnDestroy() {
+    this.onDestroy$.next(true);
+    this.onDestroy$.complete();
+  }
 
   async initCal(now: Date) {
     //make sure its clean
@@ -134,138 +104,33 @@ export class CalendarComponent implements OnInit, OnDestroy {
     this.currentDay = now.getDay();
     this.currentFullYear = now.getFullYear();
 
-    this.daysInMonth = this.getLastDayOfCurrentMonth(now);
+    this.daysInMonth = this.calendarService.getLastDayOfCurrentMonth(now);
 
     let monthKey = this.persistenceService.getMonthKey(now);
     this.persistenceData = await this.persistenceService.getFromPersistence(
       monthKey
     );
 
-    this.currentCalenderMonthArray = this.getCurrentCalenderMonthArray(
+    this.currentCalenderMonthArray = this.calendarService.getCurrentCalenderMonthArray(
       now,
       this.persistenceData || []
     );
-
-
   }
 
+  
   /**
-   * gets day of week of first day of month so it can be padded
-   * is 0 BASED
-   * 0 = sunday
-   * @param date
-   * @returns
+   * depping to inline form, using for documentation UI for now.
+   * @param reminder 
    */
-  getDayOfWeekOfFirstDay(date: Date) {
-    return moment(date).startOf('month').toDate().getDay();
+   openReminderForm(reminder?: Reminder) {
+    this.matDialog.open(ReminderFormComponent, {
+      data: {
+        reminder,
+      },
+    });
   }
 
-  /**
-   * gets number of days in month
-   * is 1 BASED
-   */
-  getLastDayOfLastMonth(date: Date) {
-    return moment(date).startOf('month').subtract(1, 'd').toDate().getDate();
-  }
-
-  getLastDayOfCurrentMonth(date: Date) {
-    return moment(date).endOf('month').toDate().getDate();
-  }
-
-  /**
-   * pass in a date and get a data representation
-   * starting on sunday like a calendar would display, including
-   * before start of current month and after as well
-   * so its easy to do a repeat and show the month in the ui.
-   * [29, 31, 1, 2, -- 30, 1, 2]
-   * @param dateToMakeArr
-   */
-  getCurrentCalenderMonthArray(
-    dateToMakeArr: Date,
-    persistenceData: any
-  ): CalendarLayoutData[] {
-
-    let lastDayOfLastMonth = this.getLastDayOfLastMonth(dateToMakeArr);
-
-    let lastDayOfCurrentMonth = this.getLastDayOfCurrentMonth(dateToMakeArr);
-
-
-    let currentCalenderLayoutData: CalendarLayoutData[] = [];
-
-    let firstSunday = moment(dateToMakeArr)
-      .startOf('month')
-      .startOf('week')
-      .toDate()
-      .getDate();
-
-    //pre month, skip if month starts on sunday!
-    if (firstSunday !== 1) {
-      for (let preI = firstSunday; preI <= lastDayOfLastMonth; preI++) {
-        currentCalenderLayoutData.push({
-          day: preI,
-
-          //not needed since not clickable
-          dateType: CalendarViewDateType.Pre,
-          events: [],
-        });
-      }
-    }
-
-    //this month
-    for (let curI = 1; curI <= lastDayOfCurrentMonth; curI++) {
-      //only current is clickable
-
-      let keyI = 'day-' + curI.toString();
-
-      const today = moment(new Date()).toDate().getDate();
-      let dateTypeLocal;
-
-      const nowDate = new Date();
-      let isCurrMonth = moment(nowDate).isSame(dateToMakeArr, 'month');
-
-      if (isCurrMonth) {
-        //current month all days before today are uneditable(can set a reminder in the past?)
-        dateTypeLocal =
-          curI < today
-            ? CalendarViewDateType.CurrentMonthButPast
-            : CalendarViewDateType.Current;
-      } else if (moment(dateToMakeArr).isAfter(nowDate)) {
-        //future month
-        dateTypeLocal = CalendarViewDateType.Current;
-      } else {
-        //past month, mark as past
-        dateTypeLocal = CalendarViewDateType.Pre;
-      }
-
-      currentCalenderLayoutData.push({
-        day: curI,
-        year: dateToMakeArr.getFullYear(),
-        month: dateToMakeArr.getMonth(),
-        dateType: dateTypeLocal,
-        events:  persistenceData[keyI]?.events || [],
-      });
-    }
-
-    //ms calender just has 42 days always, it avoids shifting layout.
-    //if we dont want that we could check for 35 and truncate there.
-    const padEndLen =
-      this.constantService
-        .ANSWER_TO_THE_ULTIMATE_QUESTION_OF_LIFE_THE_UNIVERS_AND_EVERYTHING -
-      currentCalenderLayoutData.length;
-
-    //next month padding
-    for (let postI = 1; postI <= padEndLen; postI++) {
-      currentCalenderLayoutData.push({
-        day: postI,
-
-        dateType: CalendarViewDateType.Post,
-        events: [],
-      });
-    }
-
-    return currentCalenderLayoutData;
-  }
-
+ 
   /**
    * change month to the previous
    * @param date the date to subtract
@@ -365,7 +230,6 @@ export class CalendarComponent implements OnInit, OnDestroy {
 
     this.setWeatherData(daysAhead, location);
 
-
     layoutData.selected = true;
 
     this.dateSelections = {
@@ -404,6 +268,7 @@ export class CalendarComponent implements OnInit, OnDestroy {
     };
   }
 
+  //wasnt runnning fast
   deselectAllDays() {
     // this.currentCalenderMonthArray.forEach((item) => {
     //   item.selected = false;
@@ -418,28 +283,25 @@ export class CalendarComponent implements OnInit, OnDestroy {
     this.deselectAllDays();
     let { layoutData, formData, formModeData } = data;
     const persistenceData: any = this.persistenceData;
-    
-    console.log(' persistenceData ', persistenceData);
 
+    console.log(' persistenceData ', persistenceData);
 
     //if editing change a specific index with the form data
     if (formModeData.mode === 'edit') {
       //if they updated the day in the droppy down.
       if (formModeData.dayToUpdate !== -1) {
-
         //write new one
         layoutData.day = formModeData.dayToUpdate;
-       
-        //remove old event.
-        console.log(layoutData.events, '-----')
-          // remove old event.
-          await this.rmEvent(
-            calendarLayoutDayOriginal.day,
-            formModeData.evtI,
-            calendarLayoutDayOriginal,
-            persistenceData            
-          );
 
+        //remove old event.
+        console.log(layoutData.events, '-----');
+        // remove old event.
+        await this.rmEvent(
+          calendarLayoutDayOriginal.day,
+          formModeData.evtI,
+          calendarLayoutDayOriginal,
+          persistenceData
+        );
       } else {
         //change values without updating..
         layoutData.events[formModeData.evtI] = formData;
@@ -513,10 +375,7 @@ export class CalendarComponent implements OnInit, OnDestroy {
    * @param evtI
    * @param calendarLayoutDay
    */
-  editEvent(
-    evtI: number,
-    layoutData: CalendarLayoutData
-  ) {
+  editEvent(evtI: number, layoutData: CalendarLayoutData) {
     console.log('todo edit event iwth id', evtI, layoutData);
 
     this.formModeData = {
@@ -528,14 +387,10 @@ export class CalendarComponent implements OnInit, OnDestroy {
     this.formData = layoutData.events[evtI];
   }
 
-
-
   /**
    * cancel btn
    */
   cancelEvent() {
     this.resetDateSelections();
   }
-
-
 }
