+++ src/app/components/reminder-form/reminder-form.component.spec.ts
@@ -28,5 +28,4 @@ describe('ReminderFormComponent', () => {
 
   //tests moved to inline version.
 
-
 });
